\name{MOVA-package}
\alias{MOVA-package}
\alias{MOVA}
\docType{package}
\title{
a method for evaluating the pathogenicity of missense variants using AlphaFold2
\packageTitle{MOVA}
}
\description{
\packageDescription{MOVA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MOVA}
\packageIndices{MOVA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MOVA}

Maintainer: \packageMaintainer{MOVA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
